requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
lxml>=4.9.0

# Excel support
openpyxl>=3.0.10

# Scheduler
schedule>=1.2.0

# Environment variables
python-dotenv>=0.19.0

# Web framework
flask>=2.3.0
jinja2>=3.1.0

# Optional: Selenium for dynamic sites (if needed)
# selenium>=4.8.0
# webdriver-manager>=3.8.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0